Rename rebased test imports
